tidy: Move ExtraRequestSettings to game types
