+++ crates/lib/src/protocols/mod.rs
@@ -15,4 +15,4 @@ pub mod unreal2;
 /// Reference: [Server Query](https://developer.valvesoftware.com/wiki/Server_queries)
 pub mod valve;
 
-pub use types::{ExtraRequestSettings, GenericResponse, Protocol};
+pub use types::{GenericResponse, Protocol};
